back casi completo, front checkout funcional, falta crud usuario y perfil

+++ Backend/RetroKits/RetroKits/Database/Order.cs
@@ -3,7 +3,7 @@
 public class Order
 {
     public int Id { get; set; }
-    public int TotalAmount{ get; set; }
+    public float TotalAmount{ get; set; }
     public DateTime Date {  get; set; }
     public int UserId { get; set; }
     public User User { get; set; }
